[skip ci] restructure the duplication reverter
