+++ angr/analyses/decompiler/optimization_passes/duplication_reverter/duplication_reverter.py
@@ -1,1249 +1,43 @@
 from collections import defaultdict
 from copy import deepcopy
-from typing import Optional
 import logging
 from itertools import combinations
 import itertools
 
-import networkx
 import networkx as nx
 
 import ailment
-from ailment import AILBlockWalkerBase
 from ailment.block import Block
-from ailment.statement import Statement, ConditionalJump, Jump, Assignment, Return, Label
-from ailment.expression import Const, Register, Convert, BinaryOp, Expression
+from ailment.statement import ConditionalJump, Jump, Assignment, Return, Label
+from ailment.expression import Const, Register, Convert, Expression
 
-from .optimization_pass import OptimizationPass, OptimizationPassStage
-from ..block_io_finder import BlockIOFinder
-from ..goto_manager import GotoManager
-from ..region_identifier import RegionIdentifier
-from ..structuring import RecursiveStructurer, PhoenixStructurer
-from ..structuring.structurer_nodes import IncompleteSwitchCaseHeadStatement
-from ....analyses.reaching_definitions.reaching_definitions import ReachingDefinitionsAnalysis
-from ..utils import to_ail_supergraph, remove_labels
-from ....knowledge_plugins.key_definitions.atoms import MemoryLocation
-from ....utils.graph import dominates
+from .ail_merge_graph import AILMergeGraph, AILBlockSplit
+from .errors import StructuringError, SAILRSemanticError
+from .similarity import longest_ail_graph_subseq
+
+from .utils import (
+    replace_node_in_graph,
+    shared_common_conditional_dom,
+    find_block_in_successors_by_addr,
+    copy_graph_and_nodes,
+    correct_jump_targets,
+    ail_block_from_stmts,
+    deepcopy_ail_anyjump,
+)
+from angr.analyses.decompiler.counters.boolean_counter import BooleanCounter
+from ..optimization_pass import OptimizationPass, OptimizationPassStage
+from ...block_io_finder import BlockIOFinder
+from ...block_similarity import is_similar, index_of_similar_stmts, longest_ail_subseq
+from ...goto_manager import GotoManager
+from ...region_identifier import RegionIdentifier
+from ...structuring import RecursiveStructurer, PhoenixStructurer
+from .....analyses.reaching_definitions.reaching_definitions import ReachingDefinitionsAnalysis
+from ...utils import to_ail_supergraph, remove_labels, add_labels
+from .....knowledge_plugins.key_definitions.atoms import MemoryLocation
 
 l = logging.getLogger(name=__name__)
 _DEBUG = False
 l.setLevel(logging.DEBUG)
-
-#
-# Exception Types
-#
-
-
-class StructuringError(Exception):
-    """
-    These types of errors are fatal and prevent any future working from structuring in this pass
-    """
-
-    pass
-
-
-class SAILRSemanticError(Exception):
-    """
-    These types of errors may not kill the entire analysis, but they do kill the current working round.
-    """
-
-    pass
-
-
-class UnsupportedAILNodeError(SAILRSemanticError):
-    pass
-
-
-#
-# Util Classes
-#
-
-
-class ConditionBooleanWalker(AILBlockWalkerBase):
-    """
-    This class counts the number of Boolean operators an expression has.
-    In the case of: `if (a || (b && c))`, it will count 2 Boolean operators.
-
-    TODO: this entire boolean checking semantic we use needs to be removed, see how it is used for other dels needed
-    we need to replace it with a boolean variable insertion on both branches that lead to the new block
-    say we have:
-    if (A()) {
-        do_thing();
-    }
-    if (B()) {
-        do_thing():
-    }
-
-    We want to translate it to:
-    int should_do_thing = 0;
-    if (A())
-        should_do_thing = 1;
-    if (B())
-        should_do_thing = 1;
-
-    if (should_do_thing):
-        do_thing();
-
-    Although longer, this code can be optimized to look like:
-    int should_do_thing = A() || B();
-    if (should_do_thing)
-        do_thing();
-    """
-
-    def __init__(self):
-        super().__init__()
-        self.boolean_cnt = 0
-
-    def _handle_BinaryOp(
-        self, expr_idx: int, expr: "BinaryOp", stmt_idx: int, stmt: "Statement", block: Optional["Block"]
-    ):
-        if expr.op == "LogicalAnd" or expr.op == "LogicalOr":
-            self.boolean_cnt += 1
-
-        self._handle_expr(0, expr.operands[0], stmt_idx, stmt, block)
-        self._handle_expr(1, expr.operands[1], stmt_idx, stmt, block)
-
-
-class AILMergeGraph:
-    def __init__(
-        self, graph=None, original_graph=None, conditional_block=None, original_blocks=None, original_split_blocks=None
-    ):
-        self.graph = graph or nx.DiGraph()
-        self.original_graph = original_graph or nx.DiGraph()
-        self.conditional_block = conditional_block
-
-        # All the original blocks in both graphs that are going to be merged into a single
-        # graph that is self.graph
-        self.original_blocks: dict[Block, list[Block]] = original_blocks or {}
-
-        # The mapping from an original_block in original_blocks to an AILBlockSplit,
-        # which can have an up, match, and down split. This happens when only partial
-        # blocks match from the original blocks
-        self.original_split_blocks: dict[Block, list[AILBlockSplit]] = original_split_blocks or {}
-
-        # The mapping from each block in the self.graph to the original blocks that
-        # are contained in the larger CFG
-        self.merge_blocks_to_originals: dict[Block, set[Block | AILBlockSplit]] = defaultdict(set)
-        self.merge_end_to_start: dict[Block, Block] = {}
-        self.starts = []
-        self.original_ends = []
-
-    def create_conditionless_graph(self, starting_blocks: list[Block], graph_lcs):
-        # get all the original blocks (reverted from the LCS) and their split blocks.
-        # split-blocks are blocks that need to be split at some stmt index to make the two blocks
-        # equal across both graphs. At a highlevel, the first block in both matching graphs either need
-        # to be a full match or a spilt block with a None up-split (since the up-split represent a starting
-        # stmt that mismatches).
-        self.starts = starting_blocks.copy()
-        merge_base, other_base = self.starts[:]
-        for block in self.starts:
-            og_blks_bfs_order, split_blks = ail_similarity_to_orig_blocks(block, graph_lcs, self.original_graph)
-            self.original_blocks[block] = og_blks_bfs_order
-            self.original_split_blocks[block] = [
-                AILBlockSplit.from_block_lcs(og, idx, len(lcs)) for og, (lcs, idx) in split_blks.items()
-            ]
-
-        # eliminate shared blocks that are the same in original blocks
-        shared_blocks = set(self.original_blocks[merge_base]).intersection(set(self.original_blocks[other_base]))
-        for shared_block in shared_blocks:
-            for block in self.starts:
-                if shared_block not in self.original_blocks[block]:
-                    continue
-
-                self.original_blocks[block].remove(shared_block)
-                if shared_block in self.original_split_blocks[block]:
-                    self.original_split_blocks[block].remove(shared_block)
-
-        # we now know all the original blocks that will be merged, in order of BFS
-        # so let's create the graph that will be the final output merged graph
-        #
-        # we start by finding all the blocks that are about to be split, i.e., blocks that
-        # have at least one stmt mismatching from each other
-        base_to_split = {
-            split_block.original: split_block.match_split
-            for split_block in self.original_split_blocks[merge_base]
-            if split_block.match_split is not None
-        }
-
-        # we create a new graph, full of the original blocks of the base, with blocks
-        # that should be split replaced.
-        # this graph is only the initial merge_graph needed, where only the blocks
-        self.graph, update_blocks = clone_graph_replace_splits(
-            nx.subgraph(self.original_graph, self.original_blocks[merge_base]), base_to_split
-        )
-        self._update_all_split_refs(update_blocks)
-        for update_block, new_block in update_blocks.items():
-            if update_block in starting_blocks:
-                idx = self.starts.index(update_block)
-                self.starts[idx] = new_block
-        merge_base, other_base = self.starts[:]
-
-        # Add all the blocks that start the graph that may have been split
-        # (i.e., only blocks in the blocks list for this function)
-        for block in self.starts:
-            for split_block in self.original_split_blocks[block]:
-                if split_block.up_split:
-                    merge_base_split = self._find_split_block_by_original(merge_base) or merge_base
-                    merge_base_split = (
-                        merge_base_split.match_split
-                        if isinstance(merge_base_split, AILBlockSplit)
-                        else merge_base_split
-                    )
-
-                    self.graph.add_edge(split_block.up_split, merge_base_split)
-
-        return update_blocks
-
-    def add_edges_to_condition(self, conditional_block, true_target, merge_end_pairs):
-        for match_node, merge_end_pair in merge_end_pairs.items():
-            cond_copy = conditional_block.copy()
-            cond_copy.statements = [deepcopy_ail_anyjump(cond_copy.statements[0])]
-            cond_copy.idx += 1
-
-            # fix the condition edges
-            cond_jump_stmt: ConditionalJump = cond_copy.statements[-1]
-            if not merge_end_pair:
-                l.info(f"Encountered a conditional jump that has no successors on {self.starts}!")
-                raise SAILRSemanticError("Encountered a conditional jump that has no successors! This can be bad!")
-            elif len(merge_end_pair) == 1:
-                b0 = merge_end_pair[0]
-                b0_og = list(self.merge_blocks_to_originals[b0])[0]
-                if isinstance(b0_og, AILBlockSplit):
-                    b0_og = b0_og.original
-
-                b1_og = self._find_block_pair_in_originals(b0_og)
-                if b1_og is None:
-                    l.info(f"Encountered a conditional jump that has only 1 successor on {self.starts}!")
-                    raise SAILRSemanticError(
-                        f"Encountered a conditional jump that has only 1 successor on {self.starts}!"
-                    )
-
-                b1_og_succs = list(self.original_graph.successors(b1_og))
-                assert len(b1_og_succs) == 1
-                b1 = b1_og_succs[0]
-            else:
-                b0, b1 = merge_end_pair
-
-            if true_target == self._find_og_start_by_merge_end(b0):
-                cond_jump_stmt.true_target.value = b0.addr
-                cond_jump_stmt.false_target.value = b1.addr
-            else:
-                cond_jump_stmt.false_target.value = b0.addr
-                cond_jump_stmt.true_target.value = b1.addr
-
-            self.graph.add_edge(match_node, cond_copy)
-            self.graph.add_edge(cond_copy, b0)
-            self.graph.add_edge(cond_copy, b1)
-
-    def create_mapping_to_merge_graph(self, updated_blocks: dict[Block, Block], start_blocks: list[Block]):
-        merge_base, other_base = self.starts[:]
-
-        # collect all the normal blocks, which should be easy to pair
-        for base, other in zip(self.original_blocks[merge_base], self.original_blocks[other_base]):
-            if base in updated_blocks:
-                base = updated_blocks[base]
-            self.merge_blocks_to_originals[base] = {base, other}
-
-        # collect all the split blocks, which have new addrs, and match them to merge_base
-        for sblock in self.original_split_blocks[merge_base]:
-            attrs = ("up_split", "match_split", "down_split")
-            for attr in attrs:
-                sblock_split = getattr(sblock, attr)
-                if not sblock_split:
-                    continue
-                self.merge_blocks_to_originals[sblock_split].add(sblock)
-
-        # finally, use the info from merge_base mapping to add to the mapping for other_base
-        for sblock in self.original_split_blocks[other_base]:
-            attrs = ("up_split", "match_split", "down_split")
-            base_original = self._find_merge_block_by_original(sblock.original)
-            for attr in attrs:
-                sblock_split = getattr(sblock, attr)
-                if not sblock_split:
-                    continue
-
-                if attr != "match_split":
-                    self.merge_blocks_to_originals[sblock_split].add(sblock)
-                    continue
-
-                for base_split, base_sblocks in self.merge_blocks_to_originals.items():
-                    # this split block must have only one pair right now
-                    if len(base_sblocks) > 1:
-                        continue
-
-                    base_sblock = list(base_sblocks)[0]
-                    # the original base must match
-                    if base_sblock.original != base_original:
-                        continue
-
-                    # this split block must also be the right split (up, match, down)
-                    if base_split != getattr(base_sblock, attr):
-                        continue
-
-                    self.merge_blocks_to_originals[base_split].add(sblock)
-
-        # remove any extra block that will not be in the graph because it was split up
-        deletable_blocks = set()
-        end_pair_map = {}
-        end_pairs = set()
-        merge_to_end_pair = {}
-        for split, sblocks in self.merge_blocks_to_originals.items():
-            for sblock in sblocks:
-                if isinstance(sblock, AILBlockSplit) and sblock.original in self.merge_blocks_to_originals:
-                    deletable_blocks.add(sblock.original)
-        for del_block in deletable_blocks:
-            b0, b1 = self.merge_blocks_to_originals[del_block]
-            end_pair_map[b0] = b1
-            end_pair_map[b1] = b0
-            end_pairs.add((b0, b1))
-            merge_to_end_pair[del_block] = (b0, b1)
-            del self.merge_blocks_to_originals[del_block]
-
-        # mappings are collected, now we need to connect the merge ends of the graph to the upper parts with
-        # the conditions that blocked the original
-        merge_ends = {
-            block: split for block, split in self.merge_blocks_to_originals.items() if block not in self.graph
-        }
-        assert all(len(bs) == 1 for block, bs in self.merge_blocks_to_originals.items() if block in merge_ends)
-        self.merge_end_to_start = {merge_end: self._find_og_start_by_merge_end(merge_end) for merge_end in merge_ends}
-        merge_end_pairs = defaultdict(list)
-        for merge_end, split_blocks in merge_ends.items():
-            split_block = list(split_blocks)[0]
-            if split_block.original in self.original_blocks[merge_base]:
-                merge_end_pairs[split_block.match_split].append(merge_end)
-                continue
-
-            merge_original = end_pair_map.get(split_block.original, None) or split_block.original
-            base_split = self._find_split_block_by_original(merge_original)
-            merge_end_pairs[base_split.match_split].append(merge_end)
-
-        for pair in end_pairs:
-            for block in pair:
-                self.original_ends.append(block)
-        # the case of single block merges
-        if not self.original_ends:
-            self.original_ends = start_blocks
-
-        # moved here
-        for unsplit_block, pair in merge_to_end_pair.items():
-            for block in pair:
-                other_block = pair[0] if pair[1] is block else pair[1]
-                while True:
-                    for merge, originals in self.merge_blocks_to_originals.items():
-                        if self._block_in_originals(merge, other_block):
-                            continue
-
-                        if len(originals) == 1:
-                            og = list(originals)[0]
-                            if isinstance(og, Block) and og == block:
-                                self.merge_blocks_to_originals[merge].add(other_block)
-                                break
-
-                        found = False
-                        for og in originals:
-                            if isinstance(og, AILBlockSplit) and og.original == block and merge == og.match_split:
-                                self.merge_blocks_to_originals[merge].add(other_block)
-                                found = True
-                                break
-
-                        if found:
-                            break
-
-                    else:
-                        break
-
-        return merge_end_pairs
-
-    def merged_is_split_type(self, merge_block: Block, split_type: str):
-        if split_type not in ["up_split", "down_split", "match_split"]:
-            raise Exception("Can't call like this!")
-
-        for oblock in self.merge_blocks_to_originals[merge_block]:
-            if not isinstance(oblock, AILBlockSplit):
-                continue
-
-            if getattr(oblock, split_type) == merge_block:
-                return True
-        else:
-            return False
-
-    #
-    # Private Functions
-    #
-
-    def _find_block_pair_in_originals(self, block: Block):
-        for merge, originals in self.merge_blocks_to_originals.items():
-            # need at least 2 for a pair
-            if len(originals) < 2:
-                continue
-
-            # the block we are searching for a pair for should exist in the originals
-            for og in originals:
-                if isinstance(og, Block) and og == block:
-                    break
-                elif isinstance(og, AILBlockSplit) and og.original == block:
-                    break
-            else:
-                return None
-
-            # we now know that our target block is in this originals set
-            # now we just need to find any other block that is not itself
-            for other_block in originals:
-                if isinstance(other_block, Block) and other_block != block:
-                    return other_block
-                elif isinstance(other_block, AILBlockSplit) and other_block.original != block:
-                    return other_block.original
-
-        return None
-
-    def _block_in_any_blocks(self, block: Block, any_blocks):
-        for any_block in any_blocks:
-            if block == any_block:
-                return True
-
-            if isinstance(block, AILBlockSplit) and (
-                block.up_split == block or block.match_split == block or block.down_split == block
-            ):
-                return True
-
-        return False
-
-    def _block_in_originals(self, merge_block: Block, target_block: Block):
-        for original in self.merge_blocks_to_originals[merge_block]:
-            if isinstance(original, Block):
-                if original == target_block:
-                    return True
-            elif isinstance(original, AILBlockSplit):
-                if original.original == target_block:
-                    return True
-
-        return False
-
-    def _update_all_split_refs(self, update_map: dict[Block, Block]):
-        for original, updated in update_map.items():
-            for k in list(self.original_split_blocks.keys()):
-                if k == original:
-                    self.original_split_blocks[updated] = self.original_split_blocks[k]
-                    del self.original_split_blocks[k]
-
-            for k, v in self.original_split_blocks.items():
-                for sblock in v:
-                    for attr in ["up_split", "match_split", "down_split"]:
-                        if getattr(sblock, attr) == original:
-                            setattr(sblock, attr, updated)
-
-            for k in list(self.original_blocks.keys()):
-                if k == original:
-                    self.original_blocks[updated] = self.original_blocks[k]
-                    del self.original_blocks[k]
-
-    def _find_merge_block_by_original(self, block: Block):
-        for merge_block, originals in self.merge_blocks_to_originals.items():
-            for og in originals:
-                if isinstance(og, AILBlockSplit) and og.original == block:
-                    return merge_block
-
-                if isinstance(og, Block) and og == block:
-                    return merge_block
-        else:
-            l.warning(f"Error in finding the merge block from the original block on {block}")
-            return None
-
-    def _find_split_block_by_original(self, block: Block):
-        for _, split_blocks in self.original_split_blocks.items():
-            for split_block in split_blocks:
-                if split_block.original == block:
-                    return split_block
-        else:
-            l.warning(f"Error in finding split block by original on {block}")
-            return None
-
-    def _find_og_start_by_merge_end(self, merge_end: Block):
-        original = list(self.merge_blocks_to_originals[merge_end])[0]
-        if isinstance(original, AILBlockSplit):
-            original = original.original
-
-        for og_start, og_blocks in self.original_blocks.items():
-            if original in og_blocks:
-                return og_start
-
-
-class AILBlockSplit:
-    def __init__(self, original=None, up_split=None, match_split=None, down_split=None):
-        """
-
-        :param original:
-        :param up_split:   The block split above the matched LCS
-        :param match_split:   The block containing only the matched LCS
-        :param down_split:  The block split below the matched LCS
-        """
-        self.original = original
-        self.up_split = up_split
-        self.match_split = match_split
-        self.down_split = down_split
-
-    @classmethod
-    def from_block_lcs(cls, original_block: Block, idx, len_):
-        pre, mid, post = cls.split_ail_block(original_block, idx, len_)
-        return cls(
-            original=original_block,
-            up_split=pre,
-            match_split=mid,
-            down_split=post,
-        )
-
-    @staticmethod
-    def split_ail_block(block, split_idx, split_len) -> tuple[Block | None, Block | None, Block | None]:
-        if split_len == len(block.statements):
-            return None, block, None
-
-        up_split = ail_block_from_stmts(block.statements[:split_idx], block_addr=block.addr)
-        match_split = ail_block_from_stmts(
-            block.statements[split_idx : split_idx + split_len], block_addr=None if up_split else block.addr
-        )
-        down_split = ail_block_from_stmts(block.statements[split_idx + split_len :])
-
-        return up_split, match_split, down_split
-
-    def __str__(self):
-        return f"<AILBlockSplit: OG: {self.original.__repr__()} | Up: {self.up_split.__repr__()} | Match: {self.match_split.__repr__()} | Down: {self.down_split.__repr__()}>"
-
-    def __repr__(self):
-        return self.__str__()
-
-
-#
-# Graph Utils
-#
-
-
-def add_labels(graph: nx.DiGraph):
-    new_graph = nx.DiGraph()
-    nodes_map = {}
-    for node in graph:
-        lbl = ailment.Stmt.Label(None, f"LABEL_{node.addr:x}", node.addr, block_idx=node.idx)
-        node_copy = node.copy()
-        node_copy.statements = [lbl] + node_copy.statements
-        nodes_map[node] = node_copy
-
-    new_graph.add_nodes_from(nodes_map.values())
-    for src, dst in graph.edges:
-        new_graph.add_edge(nodes_map[src], nodes_map[dst])
-
-    return new_graph
-
-
-def remove_useless_gotos(graph: nx.DiGraph):
-    new_graph = nx.DiGraph()
-    nodes_map = {}
-    for node in graph:
-        node_copy = node.copy()
-        # remove Jumps from everything except the last node in the statements list
-        node_copy.statements = [
-            stmt for stmt in node_copy.statements[:-1] if not isinstance(stmt, Jump)
-        ] + node_copy.statements[-1:]
-        nodes_map[node] = node_copy
-
-    new_graph.add_nodes_from(nodes_map.values())
-    for src, dst in graph.edges:
-        new_graph.add_edge(nodes_map[src], nodes_map[dst])
-
-    return new_graph
-
-
-def find_block_by_similarity(block, graph, node_list=None):
-    nodes = node_list if node_list else list(graph.nodes())
-    similar_blocks = []
-    for other_block in nodes:
-        if similar(block, other_block, graph=graph):
-            similar_blocks.append(other_block)
-
-    if len(similar_blocks) > 1:
-        l.warning("found multiple similar blocks")
-
-    return similar_blocks[0]
-
-
-def find_block_in_successors_by_addr(addr: int, block: ailment.Block, graph: nx.DiGraph):
-    for succ in graph.successors(block):
-        if succ.addr == addr or succ.statements[0].ins_addr == addr:
-            return succ
-    else:
-        return None
-
-
-def find_block_by_addr(graph: networkx.DiGraph, addr, insn_addr=False):
-    if insn_addr:
-
-        def _get_addr(b):
-            return b.statements[0].ins_addr
-
-    else:
-
-        def _get_addr(b):
-            return b.addr
-
-    for block in graph.nodes():
-        if _get_addr(block) == addr:
-            break
-    else:
-        block = None
-        raise Exception("The block is not in the graph!")
-
-    return block
-
-
-def clone_graph_replace_splits(graph: nx.DiGraph, split_map: dict[Block, Block]):
-    # do a deepcopy, so we don't edit the original graph
-    graph = copy_graph_and_nodes(graph)
-
-    # replace every block that has been split
-    updated_blocks = {}
-    for original_node, new_node in split_map.items():
-        graph.add_node(new_node)
-        # correct every in_edge to this node, with new targets for jumps
-        for pred in list(graph.predecessors(original_node)):
-            new_pred = pred.copy()
-            new_pred.statements[-1] = correct_jump_targets(
-                new_pred.statements[-1], {original_node.addr: new_node.addr}, new_stmt=True
-            )
-            updated_blocks[pred] = new_pred
-            replace_node_in_graph(graph, pred, new_pred)
-            graph.add_edge(new_pred, new_node)
-
-        # re-add every out_edge
-        for succ in graph.successors(original_node):
-            graph.add_edge(new_node, succ)
-
-        # finally, kill the original
-        if original_node in graph:
-            graph.remove_node(original_node)
-
-    return graph, updated_blocks
-
-
-def clone_graph_with_splits(graph_param: nx.DiGraph, split_map_param):
-    split_map = {(block.addr, block.idx): new_node for block, new_node in split_map_param.items()}
-    graph = copy_graph_and_nodes(graph_param)
-    # this loop will continue to iterate until there are no more
-    # nodes found in the split_map to change
-    while True:
-        for node in graph.nodes():
-            try:
-                new_node = split_map[(node.addr, node.idx)]
-                del split_map[(node.addr, node.idx)]
-            except KeyError:
-                continue
-
-            break
-        else:
-            break
-
-        graph.add_node(new_node)
-        for pred in graph.predecessors(node):
-            last_stmt = pred.statements[-1]
-            pred.statements[-1] = correct_jump_targets(last_stmt, {node.addr: new_node.addr}, new_stmt=True)
-
-            graph.add_edge(pred, new_node)
-
-        for succ in graph.successors(node):
-            graph.add_edge(new_node, succ)
-
-        graph.remove_node(node)
-
-    for node in graph.nodes():
-        last_stmt = node.statements[-1]
-        node.statements[-1] = correct_jump_targets(
-            last_stmt, {orig_addr: new.addr for orig_addr, new in split_map.items()}, new_stmt=True
-        )
-
-    return graph
-
-
-def replace_node_in_graph(graph: networkx.DiGraph, node, replace_with):
-    in_edges = list(graph.in_edges(node))
-    out_edges = list(graph.out_edges(node))
-
-    graph.remove_node(node)
-    graph.add_node(replace_with)
-
-    for src, _ in in_edges:
-        if src is node:
-            graph.add_edge(replace_with, replace_with)
-        else:
-            graph.add_edge(src, replace_with)
-
-    for _, dst in out_edges:
-        if dst is node:
-            graph.add_edge(replace_with, replace_with)
-        else:
-            graph.add_edge(replace_with, dst)
-
-    assert node not in graph
-
-
-def bfs_list_blocks(start_block: Block, graph: nx.DiGraph):
-    blocks = []
-    bfs = list(nx.bfs_successors(graph, start_block, depth_limit=10))
-    for blk_tree in bfs:
-        source, children = blk_tree
-        last_src_stmt = source.statements[-1] if source.statements else None
-        if (
-            last_src_stmt is None
-            or not isinstance(last_src_stmt, Statement)
-            or isinstance(last_src_stmt, IncompleteSwitchCaseHeadStatement)
-        ):
-            raise UnsupportedAILNodeError(f"Stmt {last_src_stmt} is unsupported")
-
-        if len(children) == 1:
-            blocks += children
-        elif len(children) == 2:
-            if_stmt: ConditionalJump = source.statements[-1]
-            if children[0].addr == if_stmt.true_target.value:
-                blocks += [children[0], children[1]]
-            else:
-                blocks += [children[1], children[0]]
-
-    blocks = [start_block] + blocks
-    return blocks
-
-
-def copy_graph_and_nodes(graph: nx.DiGraph, new_idx=False):
-    """
-    TODO: the function below this that does the same need to be deprecated
-    """
-    new_graph = nx.DiGraph()
-    nodes_map = {}
-    for node in graph.nodes:
-        node_copy = node.copy()
-        node_copy.statements = [stmt for stmt in node_copy.statements]
-        if new_idx:
-            node_copy.idx = node_copy.idx + 1 if isinstance(node_copy.idx, int) else 1
-        nodes_map[node] = node_copy
-
-    new_graph.add_nodes_from(nodes_map.values())
-    for src, dst in graph.edges:
-        new_graph.add_edge(nodes_map[src], nodes_map[dst])
-
-    return new_graph
-
-
-def shared_common_conditional_dom(nodes, graph: nx.DiGraph):
-    """
-    Takes n nodes and returns True only if all the nodes are dominated by the same node, which must be
-    a ConditionalJump
-
-    @param nodes:
-    @param graph:
-    @return:
-    """
-    try:
-        entry_blk = [node for node in graph.nodes if graph.in_degree(node) == 0][0]
-    except IndexError:
-        return None
-
-    idoms = nx.algorithms.immediate_dominators(graph, entry_blk)
-    """
-    ancestors = {
-        node: list(nx.ancestors(graph, node)) for node in nodes
-    }
-
-    # no node for merging can be an ancestor to the other
-    for node in nodes:
-        other_ancestors = itertools.chain.from_iterable([ances for n, ances in ancestors.items() if n != node])
-        if node in other_ancestors:
-          return None
-    """
-
-    # first check if any of the node pairs could be a dominating loop
-    b0, b1 = nodes[:]
-    if dominates(idoms, b0, b1) or dominates(idoms, b1, b0):
-        return None
-
-    node = nodes[0]
-    node_level = [node]
-    seen_nodes = set()
-    while node_level:
-        # check if any of the nodes on the current level are dominaters to all nodes
-        for cnode in node_level:
-            if not cnode.statements:
-                continue
-
-            if (
-                isinstance(cnode.statements[-1], ConditionalJump)
-                and all(dominates(idoms, cnode, node) for node in nodes)
-                and cnode not in nodes
-            ):
-                return cnode
-
-        # if no dominators found, move up a level
-        seen_nodes.update(set(node_level))
-        next_level = list(itertools.chain.from_iterable([list(graph.predecessors(cnode)) for cnode in node_level]))
-        # only add nodes we have never seen
-        node_level = set(next_level).difference(seen_nodes)
-
-    else:
-        return None
-
-
-#
-# AIL Helpers
-#
-
-
-def similar(ail_obj1, ail_obj2, graph: nx.DiGraph = None, partial=True):
-    if type(ail_obj1) is not type(ail_obj2):
-        return False
-
-    if ail_obj1 is ail_obj2:
-        return True
-
-    # AIL Blocks
-    if isinstance(ail_obj1, Block):
-        if len(ail_obj1.statements) != len(ail_obj2.statements):
-            return False
-
-        for stmt1, stmt2 in zip(ail_obj1.statements, ail_obj2.statements):
-            if not similar(stmt1, stmt2, graph=graph):
-                return False
-        else:
-            return True
-
-    # AIL Statements
-    elif isinstance(ail_obj1, Statement):
-        # if all(barr in [0x404530, 0x404573] for barr in [ail_obj1.ins_addr, ail_obj2.ins_addr]):
-        #    do a breakpoint
-
-        # ConditionalJump Handler
-        if isinstance(ail_obj1, ConditionalJump):
-            # try a simple compare
-            liked = ail_obj1.likes(ail_obj2)
-            if liked or not graph:
-                return liked
-
-            # even in partial matching, the condition must at least match
-            if not ail_obj1.condition.likes(ail_obj2.condition):
-                return False
-
-            # must use graph to know
-            for attr in ["true_target", "false_target"]:
-                t1, t2 = getattr(ail_obj1, attr).value, getattr(ail_obj2, attr).value
-                try:
-                    t1_blk, t2_blk = find_block_by_addr(graph, t1), find_block_by_addr(graph, t2)
-                except Exception:
-                    return False
-
-                # special checks for when a node is empty:
-                if not t1_blk.statements or not t2_blk.statements:
-                    # when both are empty, they are similar
-                    if len(t1_blk.statements) == len(t2_blk.statements):
-                        continue
-
-                    # TODO: implement a check for when one is empty and other is jump.
-                    # this will require a recursive call into similar() to check if a jump and empty are equal
-                    #
-                    # when one block has a jump but the other is empty, they are possibly similar
-                    # larger_blk = t1_blk if not t2_blk.statements else t2_blk
-                    # if len(larger_blk.statements) == 1 and isinstance(larger_blk.statements[-1], Jump):
-                    #    continue
-                    return False
-
-                # skip full checks when partial checking is on
-                if partial and t1_blk.statements[0].likes(t2_blk.statements[0]):
-                    continue
-
-                if not similar(t1_blk, t2_blk, graph=graph):
-                    return False
-            else:
-                return True
-
-        # Generic Statement Handler
-        else:
-            return ail_obj1.likes(ail_obj2)
-    else:
-        return False
-
-
-def ail_block_from_stmts(stmts, idx=None, block_addr=None) -> Block | None:
-    if not stmts:
-        return None
-
-    first_stmt = stmts[0]
-
-    return Block(
-        first_stmt.ins_addr if not block_addr else block_addr,
-        0,
-        statements=[stmt for stmt in stmts],
-        idx=idx or 1,
-    )
-
-
-def deepcopy_ail_jump(stmt: Jump, idx=1):
-    target: Const = stmt.target
-    tags = stmt.tags.copy()
-
-    return Jump(idx, Const(1, target.variable, target.value, target.bits, **target.tags.copy()), **tags)
-
-
-def deepcopy_ail_condjump(stmt: ConditionalJump, idx=1):
-    true_target: Const = stmt.true_target
-    false_target: Const = stmt.false_target
-    tags = stmt.tags.copy()
-
-    return ConditionalJump(
-        idx,
-        stmt.condition.copy(),
-        Const(1, true_target.variable, true_target.value, true_target.bits, **true_target.tags.copy()),
-        Const(1, false_target.variable, false_target.value, false_target.bits, **false_target.tags.copy()),
-        **tags,
-    )
-
-
-def deepcopy_ail_anyjump(stmt: Jump | ConditionalJump, idx=1):
-    if isinstance(stmt, Jump):
-        return deepcopy_ail_jump(stmt, idx=idx)
-    elif isinstance(stmt, ConditionalJump):
-        return deepcopy_ail_condjump(stmt, idx=idx)
-    else:
-        raise Exception(
-            "Attempting to deepcopy non-jump stmt, likely happen to a "
-            "block ending in no jump. Place a jump there to fix it."
-        )
-
-
-def correct_jump_targets(stmt, replacement_map: dict[int, int], new_stmt=True):
-    if not replacement_map or not isinstance(stmt, Statement):
-        return stmt
-
-    if isinstance(stmt, ConditionalJump):
-        cond_stmt = deepcopy_ail_condjump(stmt) if new_stmt else stmt
-        true_target, false_target = cond_stmt.true_target, cond_stmt.false_target
-
-        if isinstance(true_target, Const) and true_target.value in replacement_map:
-            true_target.value = replacement_map[true_target.value]
-
-        if isinstance(false_target, Const) and false_target.value in replacement_map:
-            false_target.value = replacement_map[false_target.value]
-
-        return cond_stmt
-    elif isinstance(stmt, Jump) and isinstance(stmt.target, Const):
-        jump_stmt = deepcopy_ail_jump(stmt) if new_stmt else stmt
-        target = jump_stmt.target
-
-        if isinstance(target, Const) and target.value in replacement_map:
-            target.value = replacement_map[target.value]
-
-        return jump_stmt
-    else:
-        return stmt
-
-
-#
-# Longest Common Substring Search Helpers/Functions
-#
-
-
-def _kmp_search_ail_obj(search_pattern, stmt_seq, graph=None, partial=True):
-    """
-    Uses the Knuth-Morris-Pratt algorithm for searching.
-    Found: https://code.activestate.com/recipes/117214/.
-
-    Returns a generator of positions, which will be empty if its not found.
-    """
-    # allow indexing into pattern and protect against change during yield
-    search_pattern = list(search_pattern)
-
-    # build table of shift amounts
-    shifts = [1] * (len(search_pattern) + 1)
-    shift = 1
-    for pos in range(len(search_pattern)):
-        while shift <= pos and not similar(
-            search_pattern[pos], search_pattern[pos - shift], graph=graph, partial=partial
-        ):
-            shift += shifts[pos - shift]
-        shifts[pos + 1] = shift
-
-    # do the actual search
-    start_pos = 0
-    match_len = 0
-    for c in stmt_seq:
-        while (
-            match_len == len(search_pattern)
-            or match_len >= 0
-            and not similar(search_pattern[match_len], c, graph=graph, partial=partial)
-        ):
-            start_pos += shifts[match_len]
-            match_len -= shifts[match_len]
-        match_len += 1
-        if match_len == len(search_pattern):
-            yield start_pos
-
-
-def stmts_pos_in_other(stmts, other, graph=None, all_positions=False):
-    """
-    Equivalent to asking:
-    stmts in other
-
-    @return: None or int (position start in other)
-    """
-    positions = list(_kmp_search_ail_obj(stmts, other, graph=graph))
-
-    if len(positions) == 0:
-        return None
-
-    return positions.pop() if not all_positions else positions
-
-
-def stmts_in_other(stmts, other, graph=None):
-    """
-    Returns True if the stmts (a list of Statement) is found as a subsequence in other
-
-    @return:
-    """
-
-    if stmts_pos_in_other(stmts, other, graph=graph) is not None:
-        return True
-
-    return False
-
-
-def longest_ail_subseq(stmts_list, graph=None):
-    """
-    Returns the LCS (a list of Statement) of the list of stmts (list of Statement).
-    Returns LCS, [LCS_POS_IN_0, LCS_POS_IN_1, ..., LCS_POS_IN_N]
-
-    @param stmts_list:
-    @param graph:
-    @return:
-    """
-
-    # find the longest sequence in all stmts
-    subseq = []
-    if len(stmts_list) <= 1:
-        return stmts_list[0], 0
-
-    if len(stmts_list[0]) > 0:
-        for i in range(len(stmts_list[0])):
-            for j in range(len(stmts_list[0]) - i + 1):
-                if j > len(subseq) and all(
-                    stmts_in_other(stmts_list[0][i : i + j], stmts, graph=graph) for stmts in stmts_list
-                ):
-                    subseq = stmts_list[0][i : i + j]
-
-    if not subseq:
-        return None, [None] * len(stmts_list)
-
-    return subseq, [stmts_pos_in_other(subseq, stmts, graph=graph) for stmts in stmts_list]
-
-
-def longest_ail_graph_subseq(block_list, graph):
-    # generate a graph similarity for each pair in the provided blocks
-    all_sims = [ail_graph_similarity(pair[0], pair[1], graph) for pair in itertools.combinations(block_list, 2)]
-
-    lcs, _ = longest_ail_subseq(all_sims, graph=graph)
-    return lcs
-
-
-def ail_graph_similarity(block0: Block, block1: Block, graph: nx.DiGraph, only_blocks=False):
-    b0_blocks = bfs_list_blocks(block0, graph)
-    b1_blocks = bfs_list_blocks(block1, graph)
-    similarity = []
-
-    discontinuity_blocks = set()
-    for i, b0 in enumerate(b0_blocks):
-        # getting to a block with no matching index is grounds to stop cmp
-        try:
-            b1 = b1_blocks[i]
-        except IndexError:
-            break
-
-        # if a block in the chain before did not end in LCS, don't follow it.
-        if b0 in discontinuity_blocks or b1 in discontinuity_blocks:
-            continue
-
-        # SPECIAL CASE: 1
-        # ┌─┐         ┌─┐
-        # │A├─┐     ┌─┤A│
-        # └─┘ │     │ └─┘
-        #     ▼     ▼
-        #    ┌──┐ ┌──┐
-        #    │C'│ │C'│
-        #    └┬─┘ └┬─┘
-        #     │    │
-        #     ▼    ▼
-        #       ...
-        #
-        # Both similar blocks end in a block that is actually the same block. In this case, we don't
-        # want to count this since we will create a N blocks long similarity that continues from C' all
-        # the way to the end of the graph. This is only true if C' is not the end block of the graph.
-        if b0 is b1:
-            preds = list(graph.predecessors(b0))
-            all_match = True
-            for pred1 in preds:
-                for pred2 in preds:
-                    if set(list(graph.successors(pred1))) != set(list(graph.successors(pred2))):
-                        all_match = False
-                        break
-
-                if not all_match:
-                    break
-
-            # CASE 1 confirmed, all edges in look the same. Now we check a special subset to see if the
-            # matching node has any edges out. If this node has no edges out, we are no longer ture for 1A
-            if all_match and len(list(graph.successors(b0))) != 0:
-                continue
-
-        lcs, lcs_idxs = longest_ail_subseq([b0.statements, b1.statements], graph=graph)
-        if not lcs:
-            break
-
-        # verify all the blocks end in that statement or exclude its children
-        for idx, b in enumerate([b0, b1]):
-            if len(lcs) + lcs_idxs[idx] != len(b.statements):
-                discontinuity_blocks.update(set(list(graph.successors(b))))
-
-        # can output blocks only if needed
-        similarity += lcs if not only_blocks else [(b0, b1)]
-
-    return similarity
-
-
-def ail_similarity_to_orig_blocks(orig_block, graph_similarity, graph):
-    traversal_blocks = bfs_list_blocks(orig_block, graph)
-
-    graph_sim = graph_similarity.copy()
-    orig_blocks = []
-    split_blocks = {}  # [block] = (lcs, idx)
-    for block in traversal_blocks:
-        if not graph_sim:
-            break
-
-        lcs, lcs_idxs = longest_ail_subseq([block.statements, graph_sim[: len(block.statements)]], graph=graph)
-        if block is orig_block:
-            lcs_1, lcs_idxs_1 = longest_ail_subseq([graph_sim[: len(block.statements)], block.statements], graph=graph)
-            if lcs_idxs_1[1] > lcs_idxs[0]:
-                lcs, lcs_idxs = lcs_1, lcs_idxs_1[::-1]
-
-        if not lcs:
-            break
-
-        orig_blocks.append(block)
-
-        if len(lcs) != len(block.statements):
-            split_blocks[block] = (lcs, lcs_idxs[0])
-
-        graph_sim = graph_sim[len(lcs) :]
-
-    return orig_blocks, split_blocks
-
-
-#
-# Simple Optimizations
-#
-
-
-def remove_redundant_jumps(graph: nx.DiGraph):
-    """
-    This can destroy ConditionalJumps with 2 successors but only one is a real target
-
-    @param graph:
-    @return:
-    """
-    change = False
-    while True:
-        for target_blk in graph.nodes:
-            if not target_blk.statements:
-                continue
-
-            # must end in a jump of some sort
-            last_stmt = target_blk.statements[-1]
-            if not isinstance(last_stmt, (Jump, ConditionalJump)):
-                continue
-
-            # never remove jumps that could have statements that are executed before them
-            # OR
-            # stmts that have a non-const jump (like a switch statement)
-            if isinstance(last_stmt, Jump) and (
-                len(target_blk.statements) > 1 or not isinstance(last_stmt.target, ailment.expression.Const)
-            ):
-                continue
-
-            # must have successors otherwise we could be removing a final jump out of the function
-            target_successors = list(graph.successors(target_blk))
-            if not target_successors:
-                continue
-
-            if isinstance(last_stmt, ConditionalJump):
-                if len(target_successors) > 2:
-                    continue
-
-                if len(target_successors) == 2:
-                    # skip real ConditionalJumps (ones with two different true/false target)
-                    if last_stmt.true_target.value != last_stmt.false_target.value:
-                        continue
-                    # XXX: removed for now
-                    # two successors, verify one is outdated and one matches the true_target value
-                    # which gauntnees we can remove the other
-                    if last_stmt.true_target.value not in [succ.addr for succ in target_successors]:
-                        continue
-
-                    # remove the outdated successor edge
-                    outdated_blks = [blk for blk in target_successors if blk.addr != last_stmt.true_target.value]
-                    if len(outdated_blks) != 1:
-                        continue
-
-                    outdated_blk = outdated_blks[0]
-                    graph.remove_edge(target_blk, outdated_blk)
-                    l.info(f"Removing simple redundant jump/cond: {(target_blk, outdated_blk)}")
-                    # restart the search because we fixed edges
-                    change |= True
-                    break
-
-            # At this point we have two situation:
-            # - a single stmt node ending in a Jump
-            # - a (possibly) multi-stmt node ending in a conditional Jump w/ 1 successor
-            successor = target_successors[0]
-
-            # In the case of the conditional jump with multiple statements before the jump, we just transfer this
-            # block's statements over to the next block, excluding the jump
-            if len(target_blk.statements) > 1:
-                successor.statements = target_blk.statements[:-1] + successor.statements
-
-            # All the predecessors need to now point to the successor of the node that is about to be removed
-            for pred in graph.predecessors(target_blk):
-                last_pred_stmt = pred.statements[-1]
-                pred.statements[-1] = correct_jump_targets(last_pred_stmt, {target_blk.addr: successor.addr})
-                graph.add_edge(pred, successor)
-
-            graph.remove_node(target_blk)
-            l.debug(f"removing node in simple redundant: {target_blk}")
-            change |= True
-            break
-        else:
-            # Finishing the loop without every breaking out of the loop means we did not change
-            # anything in this iteration, which means we hit the Fixedpoint
-            break
-
-    return graph, change
-
-
-#
-# Main Analysis
-#
-
-
-def all_has_path_to(sources: list[Block], sinks: list[Block], graph):
-    for source in sources:
-        for sink in sinks:
-            has_path = False
-            try:
-                has_path = nx.has_path(graph, source, sink)
-            except Exception:
-                pass
-
-            if not has_path:
-                return False
-    return True
 
 
 class DuplicationReverter(OptimizationPass):
@@ -1314,7 +108,7 @@ class DuplicationReverter(OptimizationPass):
                     )
                     output_graph = False
 
-            self.out_graph = add_labels(remove_useless_gotos(self.out_graph)) if output_graph else None
+            self.out_graph = add_labels(self.remove_broken_jumps(self.out_graph)) if output_graph else None
 
     def deduplication_analysis(self, max_fix_attempts=30, max_guarding_conditions=10):
         self.write_graph = remove_labels(to_ail_supergraph(copy_graph_and_nodes(self._graph)))
@@ -1628,7 +422,33 @@ class DuplicationReverter(OptimizationPass):
 
     @staticmethod
     def boolean_operators_in_condition(condition: Expression):
-        walker = ConditionBooleanWalker()
+        """
+        TODO: this entire boolean checking semantic we use needs to be removed, see how it is used for other dels needed
+        we need to replace it with a boolean variable insertion on both branches that lead to the new block
+        say we have:
+        if (A()) {
+            do_thing();
+        }
+        if (B()) {
+            do_thing():
+        }
+
+        We want to translate it to:
+        int should_do_thing = 0;
+        if (A())
+            should_do_thing = 1;
+        if (B())
+            should_do_thing = 1;
+
+        if (should_do_thing):
+            do_thing();
+
+        Although longer, this code can be optimized to look like:
+        int should_do_thing = A() || B();
+        if (should_do_thing)
+            do_thing();
+        """
+        walker = BooleanCounter()
         walker.walk_expression(condition)
         return walker.boolean_cnt
 
@@ -1714,7 +534,7 @@ class DuplicationReverter(OptimizationPass):
                     if b1_unmatched in prev_moved:
                         continue
 
-                    unmatched_b2_positions = stmts_pos_in_other([b1_unmatched], b2.statements, all_positions=True)
+                    unmatched_b2_positions = index_of_similar_stmts([b1_unmatched], b2.statements, all_positions=True)
                     if unmatched_b2_positions is None:
                         continue
 
@@ -1855,7 +675,7 @@ class DuplicationReverter(OptimizationPass):
         mismatched_blocks = {}
         for target_type in block_to_target_map[block1].keys():
             t1_blk, t2_blk = block_to_target_map[block1][target_type], block_to_target_map[block2][target_type]
-            if not similar(t1_blk, t2_blk, partial=True):
+            if not is_similar(t1_blk, t2_blk, partial=True):
                 mismatched_blocks[target_type] = {block1: t1_blk, block2: t2_blk}
 
         if len(mismatched_blocks) != 1:
@@ -2186,19 +1006,19 @@ class DuplicationReverter(OptimizationPass):
                 # we must use the more expensive `similar` function to tell on the graph if they are
                 # stmts that result in the same successors
                 try:
-                    is_similar = similar(b0, b1, graph=self.read_graph)
+                    stmt_is_similar = is_similar(b0, b1, graph=self.read_graph)
                 except Exception:
                     continue
 
                 # Case 2:
                 # [if(a)] == [if(a)]
                 # and at least one child for the correct target type matches
-                if not is_similar:
+                if not stmt_is_similar:
                     # TODO: fix this and add it back
                     # is_similar = self.similar_conditional_when_single_corrected(b0, b1, self.write_graph)
                     pass
 
-                if is_similar:
+                if stmt_is_similar:
                     initial_candidates.append((b0, b1))
                     continue
 
@@ -2223,7 +1043,7 @@ class DuplicationReverter(OptimizationPass):
 
                 for stmt1 in b1.statements:
                     # XXX: used to be just likes()
-                    if similar(stmt0, stmt1, graph=self.write_graph):
+                    if is_similar(stmt0, stmt1, graph=self.write_graph):
                         stmt_in_common = True
                         break
 
@@ -2386,11 +1206,11 @@ class DuplicationReverter(OptimizationPass):
                 # special case: when we only have a single stmt
                 if len(b0.statements) == len(b1.statements) == 1:
                     try:
-                        is_similar = similar(b0, b1, graph=graph)
+                        stmt_is_similar = is_similar(b0, b1, graph=graph)
                     except Exception:
                         continue
 
-                    if not is_similar:
+                    if not stmt_is_similar:
                         continue
 
                 elif not b0.likes(b1):
@@ -2421,6 +1241,112 @@ class DuplicationReverter(OptimizationPass):
 
         return graph, change
 
+    @staticmethod
+    def remove_redundant_jumps(graph: nx.DiGraph):
+        """
+        This can destroy ConditionalJumps with 2 successors but only one is a real target
+
+        @param graph:
+        @return:
+        """
+        change = False
+        while True:
+            for target_blk in graph.nodes:
+                if not target_blk.statements:
+                    continue
+
+                # must end in a jump of some sort
+                last_stmt = target_blk.statements[-1]
+                if not isinstance(last_stmt, (Jump, ConditionalJump)):
+                    continue
+
+                # never remove jumps that could have statements that are executed before them
+                # OR
+                # stmts that have a non-const jump (like a switch statement)
+                if isinstance(last_stmt, Jump) and (
+                    len(target_blk.statements) > 1 or not isinstance(last_stmt.target, ailment.expression.Const)
+                ):
+                    continue
+
+                # must have successors otherwise we could be removing a final jump out of the function
+                target_successors = list(graph.successors(target_blk))
+                if not target_successors:
+                    continue
+
+                if isinstance(last_stmt, ConditionalJump):
+                    if len(target_successors) > 2:
+                        continue
+
+                    if len(target_successors) == 2:
+                        # skip real ConditionalJumps (ones with two different true/false target)
+                        if last_stmt.true_target.value != last_stmt.false_target.value:
+                            continue
+                        # XXX: removed for now
+                        # two successors, verify one is outdated and one matches the true_target value
+                        # which gauntnees we can remove the other
+                        if last_stmt.true_target.value not in [succ.addr for succ in target_successors]:
+                            continue
+
+                        # remove the outdated successor edge
+                        outdated_blks = [blk for blk in target_successors if blk.addr != last_stmt.true_target.value]
+                        if len(outdated_blks) != 1:
+                            continue
+
+                        outdated_blk = outdated_blks[0]
+                        graph.remove_edge(target_blk, outdated_blk)
+                        l.info(f"Removing simple redundant jump/cond: {(target_blk, outdated_blk)}")
+                        # restart the search because we fixed edges
+                        change |= True
+                        break
+
+                # At this point we have two situation:
+                # - a single stmt node ending in a Jump
+                # - a (possibly) multi-stmt node ending in a conditional Jump w/ 1 successor
+                successor = target_successors[0]
+
+                # In the case of the conditional jump with multiple statements before the jump, we just transfer this
+                # block's statements over to the next block, excluding the jump
+                if len(target_blk.statements) > 1:
+                    successor.statements = target_blk.statements[:-1] + successor.statements
+
+                # All the predecessors need to now point to the successor of the node that is about to be removed
+                for pred in graph.predecessors(target_blk):
+                    last_pred_stmt = pred.statements[-1]
+                    pred.statements[-1] = correct_jump_targets(last_pred_stmt, {target_blk.addr: successor.addr})
+                    graph.add_edge(pred, successor)
+
+                graph.remove_node(target_blk)
+                l.debug(f"removing node in simple redundant: {target_blk}")
+                change |= True
+                break
+            else:
+                # Finishing the loop without every breaking out of the loop means we did not change
+                # anything in this iteration, which means we hit the Fixedpoint
+                break
+
+        return graph, change
+
+    @staticmethod
+    def remove_broken_jumps(graph: nx.DiGraph):
+        """
+        Removes jumps found in the middle of nodes from merging them into a single node
+        """
+        new_graph = nx.DiGraph()
+        nodes_map = {}
+        for node in graph:
+            node_copy = node.copy()
+            # remove Jumps from everything except the last node in the statements list
+            node_copy.statements = [
+                stmt for stmt in node_copy.statements[:-1] if not isinstance(stmt, Jump)
+            ] + node_copy.statements[-1:]
+            nodes_map[node] = node_copy
+
+        new_graph.add_nodes_from(nodes_map.values())
+        for src, dst in graph.edges:
+            new_graph.add_edge(nodes_map[src], nodes_map[dst])
+
+        return new_graph
+
     def simple_optimize_graph(self, graph):
         def _to_ail_supergraph(graph_):
             # make supergraph conversion always say no change
@@ -2428,7 +1354,7 @@ class DuplicationReverter(OptimizationPass):
 
         new_graph = graph.copy()
         opts = [
-            remove_redundant_jumps,
+            self.remove_redundant_jumps,
             _to_ail_supergraph,
         ]
