+++ angr/analyses/decompiler/optimization_passes/duplication_reverter/errors.py
@@ -1,0 +1,18 @@
+class StructuringError(Exception):
+    """
+    These types of errors are fatal and prevent any future working from structuring in this pass
+    """
+
+    pass
+
+
+class SAILRSemanticError(Exception):
+    """
+    These types of errors may not kill the entire analysis, but they do kill the current working round.
+    """
+
+    pass
+
+
+class UnsupportedAILNodeError(SAILRSemanticError):
+    pass
